[feat] Add authorization

+++ src/pages/login/lib/use-handle-logout.ts
@@ -1,10 +1,14 @@
 import { useCallback } from 'react'
-import { useAuthContext } from 'processes/auth/lib'
+import {useUserAuthContext} from "features/user-auth-context/lib";
 
 export const useHandleLogout = () => {
-  const { setIsAuthorized } = useAuthContext()
+  const setIsAuthorized = useUserAuthContext((contextValue) => contextValue.setIsAuthorized)
 
   const handleLogout = useCallback(() => {
+    if (!setIsAuthorized) {
+      return
+    }
+
     localStorage.clear()
     setIsAuthorized(false)
   }, [setIsAuthorized])
